Calculate recombination probabilities

Calculate the recombination probabilities, rho, from a given recombination map for the haplotype data currently in the kalis cache.

Usage

CalcRho(cM = 0, s = 1, gamma = 1, floor = TRUE)

Arguments

cM

a vector specifying the recombination distance between variants in centimorgans. Note element i of this vector should be the distance between variants i and i+1 (not i and i-1), and thus length one less than the number of variants. This can be easily obtained by applying diff() to a recombination map 'CDF'. By default, recombination probabilities are zero, meaning that all variants are perfectly linked.

s

a scalar multiplier on the recombination map (related to effective population size).

gamma

a scalar power to which the Morgan distances are raised.

floor

if TRUE (default) any recombination probabilities below machine precision (1e-16) will be zeroed out. If FALSE raw recombination probabilities will be preserved.

Value

A vector of recombination probabilities which can be used as the rho argument to the Parameters() function when creating a parameter set.

Details

This is a utility function to calculate the recombination probabilities rho (the Li and Stephens hidden Markov model transition probabilities) from a recombination map/distances.

NOTE: the corresponding haplotype data must have already been inserted into the kalis cache by a call to CacheHaplotypes(), since this function performs checks to confirm the dimensionality matches.

TODO: add kalis paper cross ref. See page 3 in Supplemental Information for the original ChromoPainter paper (Lawson et al., 2012) for motivation behind our parameterisation, which is as follows:

$$\rho = \exp(-s \times cM^\gamma) - 1$$

References

Lawson, D. J., Hellenthal, G., Myers, S., & Falush, D. (2012). Inference of population structure using dense haplotype data. PLoS genetics, 8(1).

See also

Examples

# Load the mini example data and recombination map from the package built-in #' # dataset
data("SmallHaps")
data("SmallMap")

# Load haplotypes into cache
CacheHaplotypes(SmallHaps)
#> Warning: haplotypes already cached ... overwriting existing cache.

# Compute the recombination probabilities, leaving s and gamma as default
rho <- CalcRho(diff(SmallMap))
rho
#>   [1] 7.000000e-08 1.000000e-08 5.000000e-08 3.000000e-08 6.000000e-08
#>   [6] 5.000000e-08 3.000000e-08 3.000000e-08 2.000000e-08 0.000000e+00
#>  [11] 2.000000e-08 1.000000e-08 1.300000e-07 0.000000e+00 5.000000e-08
#>  [16] 9.000000e-08 3.000000e-08 1.000000e-08 2.000000e-08 1.000000e-08
#>  [21] 4.000000e-08 3.000000e-08 1.000000e-08 6.000000e-08 5.000000e-08
#>  [26] 1.400000e-07 2.000000e-08 2.000000e-08 2.200000e-07 0.000000e+00
#>  [31] 6.000000e-08 2.000000e-08 4.000000e-08 8.000000e-08 2.000000e-08
#>  [36] 6.000000e-08 1.000000e-08 7.000000e-08 0.000000e+00 1.000000e-08
#>  [41] 1.000000e-07 5.000000e-08 7.000000e-08 1.000000e-08 4.000000e-08
#>  [46] 3.000000e-08 4.000000e-08 3.000000e-08 1.000000e-08 0.000000e+00
#>  [51] 7.000000e-08 1.000000e-08 2.000000e-08 3.000000e-08 2.000000e-08
#>  [56] 1.200000e-07 3.000000e-08 7.000000e-08 6.000000e-08 9.000000e-08
#>  [61] 1.500000e-07 8.000000e-08 1.000000e-08 0.000000e+00 1.300000e-07
#>  [66] 3.000000e-08 2.000000e-08 1.000000e-08 1.000000e-08 8.000000e-08
#>  [71] 0.000000e+00 2.000000e-08 6.000000e-08 7.000000e-08 4.000000e-08
#>  [76] 7.000000e-08 6.000000e-08 3.000000e-08 2.000000e-08 7.000000e-08
#>  [81] 0.000000e+00 2.000000e-08 7.000000e-08 3.000000e-08 0.000000e+00
#>  [86] 4.000000e-08 1.000000e-08 1.000000e-08 6.000000e-08 0.000000e+00
#>  [91] 1.000000e-08 2.000000e-08 2.000000e-08 1.000000e-08 1.000000e-08
#>  [96] 1.000000e-08 0.000000e+00 1.000000e-08 1.000000e-08 1.000000e-08
#> [101] 1.000000e-08 1.000000e-08 1.000000e-08 1.000000e-08 1.000000e-08
#> [106] 0.000000e+00 2.000000e-08 1.000000e-08 1.000000e-08 0.000000e+00
#> [111] 5.000000e-08 3.000000e-08 1.000000e-08 2.000000e-08 2.000000e-08
#> [116] 1.000000e-08 1.329999e-06 3.309995e-06 2.899996e-06 1.400000e-07
#> [121] 2.000000e-08 4.899999e-07 8.899996e-07 1.829998e-06 2.400000e-07
#> [126] 9.099996e-07 7.099997e-07 5.199999e-07 1.639999e-06 1.349999e-06
#> [131] 1.200000e-07 2.700000e-07 2.000000e-08 5.000000e-08 2.000000e-08
#> [136] 1.600000e-07 1.200000e-07 3.799999e-07 5.999998e-07 4.359990e-06
#> [141] 4.699999e-07 1.100000e-07 7.000000e-08 2.000000e-08 5.000000e-08
#> [146] 2.000000e-08 1.000000e-08 0.000000e+00 4.000000e-08 3.000000e-08
#> [151] 0.000000e+00 8.000000e-08 4.000000e-08 6.000000e-08 3.000000e-08
#> [156] 3.000000e-08 8.000000e-08 1.900000e-07 2.229998e-06 6.109981e-06
#> [161] 3.009995e-06 3.849993e-06 1.759998e-06 4.000000e-08 2.000000e-08
#> [166] 2.000000e-08 6.000000e-08 6.000000e-08 6.000000e-08 4.199999e-07
#> [171] 9.000000e-08 2.200000e-07 2.000000e-08 5.000000e-08 2.000000e-08
#> [176] 3.100000e-07 1.100000e-07 1.600000e-07 9.999999e-08 1.500000e-07
#> [181] 1.000000e-07 1.809998e-06 3.000000e-08 1.929998e-06 4.699999e-07
#> [186] 3.100000e-07 2.000000e-07 3.399999e-07 2.600000e-07 1.300000e-07
#> [191] 1.119999e-06 3.699999e-07 9.999999e-08 2.100000e-07 1.900000e-07
#> [196] 1.700000e-07 4.000000e-08 3.199999e-07 1.800000e-07 1.000000e-07
#> [201] 7.000000e-08 8.000000e-08 1.000000e-08 0.000000e+00 6.000000e-08
#> [206] 3.000000e-08 3.000000e-08 6.000000e-08 0.000000e+00 6.000000e-08
#> [211] 0.000000e+00 5.000000e-08 0.000000e+00 2.000000e-08 1.000000e-08
#> [216] 0.000000e+00 2.000000e-08 6.000000e-08 4.000000e-08 8.000000e-08
#> [221] 0.000000e+00 2.000000e-08 3.000000e-08 1.000000e-08 2.000000e-08
#> [226] 1.000000e-08 2.000000e-08 1.300000e-07 1.300000e-07 2.000000e-08
#> [231] 2.000000e-08 0.000000e+00 8.000000e-08 1.200000e-07 2.000000e-08
#> [236] 2.400000e-07 5.599998e-07 2.300000e-07 2.100000e-07 2.800000e-07
#> [241] 2.000000e-08 5.000000e-08 6.000000e-08 5.869983e-06 9.799995e-07
#> [246] 6.299998e-07 4.679989e-06 5.000000e-08 5.000000e-08 4.000000e-08
#> [251] 2.100000e-07 1.200000e-07 1.700000e-07 8.000000e-08 7.000000e-08
#> [256] 9.000000e-08 4.000000e-08 6.000000e-08 2.000000e-08 2.000000e-08
#> [261] 4.000000e-08 7.000000e-08 6.000000e-08 6.000000e-08 1.000000e-07
#> [266] 1.000000e-08 5.000000e-08 3.000000e-08 2.000000e-08 0.000000e+00
#> [271] 1.000000e-07 8.000000e-08 4.000000e-08 2.000000e-08 1.100000e-07
#> [276] 4.000000e-08 3.000000e-08 9.000000e-08 5.000000e-08 5.000000e-08
#> [281] 0.000000e+00 3.000000e-08 1.000000e-08 1.000000e-08 8.000000e-08
#> [286] 6.000000e-08 3.000000e-08 2.000000e-08 0.000000e+00 1.000000e-08
#> [291] 2.000000e-08 6.000000e-08 2.000000e-08 1.000000e-08 2.000000e-08
#> [296] 2.000000e-08 3.000000e-08 0.000000e+00 6.000000e-08 1.000000e-08
#> [301] 2.000000e-08 0.000000e+00 1.000000e-08 3.000000e-08 1.000000e-08
#> [306] 2.000000e-08 2.000000e-08 2.000000e-08 4.000000e-08 0.000000e+00
#> [311] 5.000000e-08 1.000000e-08 0.000000e+00 3.000000e-08 0.000000e+00
#> [316] 1.000000e-08 0.000000e+00 5.000000e-08 0.000000e+00 1.000000e-08
#> [321] 2.000000e-08 1.000000e-08 1.000000e-08 8.000000e-08 2.000000e-08
#> [326] 0.000000e+00 3.000000e-08 2.000000e-08 0.000000e+00 7.000000e-08
#> [331] 6.000000e-08 0.000000e+00 4.000000e-08 0.000000e+00 4.000000e-08
#> [336] 4.000000e-08 9.000000e-08 1.300000e-07 6.000000e-08 1.400000e-07
#> [341] 6.799998e-07 8.799996e-07 5.000000e-08 4.000000e-08 1.200000e-07
#> [346] 1.800000e-07 1.000000e-08 7.000000e-08 4.000000e-08 3.000000e-08
#> [351] 2.000000e-08 3.000000e-08 3.000000e-08 0.000000e+00 7.000000e-08
#> [356] 0.000000e+00 0.000000e+00 1.400000e-07 2.000000e-08 7.000000e-08
#> [361] 1.100000e-07 1.000000e-08 0.000000e+00 1.200000e-07 3.000000e-08
#> [366] 1.000000e-08 0.000000e+00 4.000000e-08 1.000000e-08 1.000000e-08
#> [371] 5.000000e-08 7.000000e-08 3.000000e-08 2.000000e-08 2.000000e-08
#> [376] 2.000000e-08 4.000000e-08 4.000000e-08 3.000000e-08 1.539999e-06
#> [381] 6.199998e-07 3.000000e-08 3.399999e-07 3.399999e-07 2.400000e-07
#> [386] 1.600000e-07 4.000000e-08 9.999999e-08 0.000000e+00 3.000000e-08
#> [391] 3.399999e-07 4.000000e-08 6.000000e-08 1.000000e-08 3.000000e-08
#> [396] 1.000000e-08 5.000000e-08 4.000000e-08 1.400000e-07