Distance matrix

Utility for calculating posterior marginal probabilities and distance matrices at, in between, or excluding variants.

DistMat(
  fwd,
  bck,
  neg.log = TRUE,
  standardize = FALSE,
  symmetrize = TRUE,
  beta.theta.opts = list(bias = 0.5)
)

Arguments

fwd
a forward table as returned by MakeForwardTable

bck
a backward table as returned by MakeBackwardTable

neg.log
a logical, should a negative logarithm be applied to the posterior copying probabilities to make them distances

standardize
a logical; should the columns be centered and scaled to have unit variance (done after neg.log transform)

symmetrize
a logical; should the final result be symmetried ( for a matrix M, 0.5*(M + t(M)) )

beta.theta.opts
a list; see Details.

Value

Matrix of distances. The \((j,i)\)-th element of of the returned matrix is the inferred distance \(d_{ji}\) between haplotypes \(j\) and \(i\) at the current locus.

Details

This computes a local probability or distance matrix based on the forward and backward tables at a certain locus. The default usage is provide forward and backward tables at the same locus \(l\). so that the \((j,i)\)-th element of of the returned matrix is the inferred distance \(d_{ji}\) between haplotypes \(j\) and \(i\) at the current locus, \(l\), of the two tables given the haplotypes observed (over the whole sequence).

In particular,

$$d_{ji} = -\frac{log(p_{ji}) + log(p_{ij})}{2}$$

where \(p_{ji}\) is the posterior marginal probability that \(j\) is coped by \(i\) at the current locus of the two tables, \(l\), given the haplotypes observed (over the whole sequence).

By convention, \(d_{ii} = 0\) for all \(i\).

A matrix of the raw posterior marginal probabilities \(p_{ji}\) may be obtained by setting neg.log = FALSE, standardize = FALSE, and symmetrize = FALSE.

This function also allows users to calculate distance matrices in between variants and also to calculate matrices that exclude a set of consecutive variants by passing a backward table in beta.theta space. If in beta.theta space, bck$l may be greater than but not equal to fwd$l. When passed a backward table in beta.theta space, the transition probabilities rho between fwd$l and bck$l will be summed together to form a total, rho_total. If bias is provided as named element in the list beta.theta.opts, then the tables will be propagated to meet in between with the given bias. More explicitly, the forward table will be propagated rho_total * bias and the backward table will be propagated a distance of rho_total * (1-bias). The distance to propagate the forward and backward tables respectively can be given by passing rho1step.fwd and rho1step.bck as arguments in beta.theta.opts, which override bias.

See also

Examples

if (FALSE) {
# To get the distances at, say, locus 10 ...
Forward(fwd, pars, 10, nthreads = 8)
Backward(bck, pars, 10, nthreads = 8)
d <- DistMat(fwd, bck)
}