Create a kalisBackwardTable

Allocates the memory for and initializes up a backward table.

MakeBackwardTable(pars, from_recipient = 1, to_recipient = Inf)

Arguments

pars
a kalisParameters environment specifying the genetics parameters to be associated with this forward table. These parameters can be set up by using the Parameters function.

from_recipient
first recipient haplotype if creating a partial forward table. By default includes from the first recipient haplotype.

to_recipient
last recipient haplotype if creating a partial forward table. By default includes to the last recipient haplotype.

Value

a specialized list of class kalisBackwardTable. For a given kalisBackwardTable, bck, bck$l denotes the current locus position of fwd. bck$beta is a matrix of rescaled backward probabilities under the Li & Stephens HMM. Each column of bck$beta corresponds to an independent HMM such that \(\beta^\ell_{ji}\) is proportional to the probability of observing haplotype \(i\) from locus \(l+1\) up through locus \(L\) given that haplotype \(j\) is copied by haplotype \(i\) at locus \(\ell\). bck$beta.g and bck$beta.g2 are both vectors containing scaling constants needed to continue propagating the HMM (please see kalis paper for details).

kalisBackwardTable also carries with it a checksum key for the parameters pars it was provided. If a user attempts to interact a kalisBackwardTable with a kalisForwardTable with mismatched parameters, an error will be thrown.

Details

MakeBackwardTable returns a kalisBackwardTable object appropriate for a given set of haplotypes (that must have already been cached by CacheAllHaplotypes) and a given set of HMM parameters specified by pars. The returned kalisBackwardTable is initialized at the end of the cached haplotypes (technically bck$l=2,147,483,647 for computational reasons) and is ready to be propagated to a given target locus with the function Backward.

Since there is an independent hidden Markov model run for each recipient haplotype, it is possible to create a partial forward table object which corresponds to a subset of recipients using the from_recipient and to_recipient arguments.

See also

Examples

# NOT RUN {
MakeBackwardTable(...)
# }