Posterior marginal probabilities

Calculated the posterior marginal probabilities at a given locus using forward and backward tables propagated to that position.

Usage

PostProbs(
  fwd,
  bck,
  unif.on.underflow = FALSE,
  M = NULL,
  beta.theta.opts = NULL,
  nthreads = min(parallel::detectCores(logical = FALSE), fwd$to_recipient -
    fwd$from_recipient + 1)
)

Arguments

fwd

a forward table as returned by MakeForwardTable

bck

a backward table as returned by MakeBackwardTable

unif.on.underflow

a logical; if TRUE, then if all probabilities in a column underflow, then they will be set to \(1/(N-1)\) instead of 0

M

an pre-existing matrix into which to write the probabilities, can yield substantial speed up but requires special attention (see Details)

beta.theta.opts

a list; see Details.

nthreads

the number of CPU cores to use. By default no parallelism is used.

from_recipient

offset for distributed problems, see kalis distributed.

Value

Matrix of posterior marginal probabilities. The \((j,i)\)-th element of of the returned matrix is the probability that \(j\) is copied by \(i\) at the current locus, \(l\), of the two tables, given the haplotypes observed (over the whole sequence).

If calculating many posterior probability matrices in succession, providing a pre-existing matrix M that can be updated in-place can drastically increase speed by eliminating the time needed for memory allocation. Be warned, since the matrix is updated in-place, if any other variables point to the same memory address, they will also be simultaneously overwritten. For example, writing

M <- matrix(0,nrow(fwd$alpha),ncol(fwd$alpha))
P <- M
PostProbs(fwd,bck,M=M)

will update M and P simultaneously.

Details

The forward and backward tables must be at the same locus in order for them to be combined to yield the posterior marginal probabilities at locus \(l\). The \((j,i)\)-th element of of the returned matrix is the probability that \(j\) is copied by \(i\) at the current locus, \(l\), of the two tables, given the haplotypes observed (over the whole sequence).

Note that each column represents an independent HMM.

By convention, every diagonal element is zero.

See also

Examples

if (FALSE) {
# To get the posterior probabilities at, say, locus 10 ...
Forward(fwd, pars, 10, nthreads = 8)
Backward(bck, pars, 10, nthreads = 8)
p <- PostProbs(fwd, bck)
}