Retrieve haplotypes from memory cache

Retrieve haplotypes from the memory cache, converting the raw binary into a simple integer 0/1 for inspection and use in R.

QueryCache(hap.ids = NA, loci.ids = NA, hap.index = NA, loci.index = NA)

Arguments

hap.ids
which haplotypes to retrieve from the cache, specified by ID. Cannot be specified at the same time as hap.index

loci.ids
which loci to retrieve from the cache, specified by ID. Cannot be specified at the same time as loci.index

hap.index
which haplotypes to retrieve from the cache, specified as a (vector) index. This enables specifying haplotypes by offset in the order they were loaded into the cache (from 1 to the number of haplotypes). Cannot be specified at the same time as hap.ids

loci.index
which loci to retrieve from the cache, specified as a (vector) index. This enables specifying loci by offset in the order they were loaded into the cache (from 1 to the number of loci). Cannot be specified at the same time as loci.ids

Value

A matrix of 0/1 integers with length rows and length(ids) columns, such that haplotypes appear in columns.

Details

To achieve higher performance, kalis internally represents haplotypes in an efficient raw binary format in memory which cannot be directly viewed or manipulated in R. This function enables you to copy whole or partial haplotypes out of this low-level format and into a standard R matrix of 0's and 1's.

See also

Examples

# Examples
if (FALSE) {
# For the purposes of an example, fill the cache with random haplotypes ...
n.haps <- 100
n.loci <- 20000
haps <- matrix(sample(0:1, n.haps*n.loci, replace = TRUE),
               nrow = n.loci,
               ncol = n.haps)
CacheHaplotypes(haps)

# ... and confirm we can read a chosen portion back.  Try to read back
# the 10th and 11th haplotypes from loci 50 to 250 inclusive
res <- QueryCache(10:11, 50, 201)
all(res == haps[50:250, 10:11])
}