Propagate a backward table

Propagates a kalisBackwardTable to an upstream locus position. The table is updated in-place.

Backward(bck, pars, t = bck$l - 1, beta.theta = FALSE, nthreads = 1)

Arguments

bck
a kalisBackwardTable object, as returned by MakeBackwardTable.

pars
a kalisParameters object, as returned by Parameters.

t
a locus position to move the backward table to. Must be less than or equal to current locus position of bck. By default, it simply advances to the locus immediately before.

beta.theta
logical indicating whether the table should be returned in beta-theta space or in the standard space upon reaching the target locus t.

nthreads
the number of CPU cores to use. By default no parallelism is used.

Value

There is nothing returned. For performance reasons, bck is updated in-place.

Details

Backward implements the backward algorithm to advance the Li and Stephens rescaled hidden Markov model backward probabilities to a new target locus.

See also

Examples

if (FALSE) {
# Assuming that pars already contains a Parameters object, we can create a
# new backward table and propagate to locus 100 using 8-core parallelsim:
bck <- MakeBackwardTable()
Backward(bck, pars, 100, nthreads = 8)
}