Calculate recombination probabilities

Calculate the recombination probabilities, rho, from a given recombination map.

CalcRho(cM = 0, s = 1, gamma = 1, floor = TRUE)

Arguments

cM
a vector specifying the recombination distance between variants in centimorgans. Note element i of this vector should be the distance between variants i and i+1 (not i and i-1), and thus length one less than the number of variants. This can be easily obtained by applying diff to a recombination map 'CDF'. By default, recombination probabilities are zero, meaning that all variants are perfectly linked.

s
a scalar multiplier on the recombination map (related to effective population size).

gamma
a scalar power to which the Morgan distances are raised.

floor
if TRUE (default) any recombination probabilities below machine precision (1e-16) will be zeroed out. If FALSE raw recombination probabilities will be preserved.

Value

A vector of recombination probabilities which can be used as the rho argument to the Parameters function when creating a parameter set.

Details

This convenient function to calculate the recombination probabilities rho (the Li and Stephens hidden Markov model transition probabilities) from a recombination map.

TODO: add kalis paper cross ref. See page 3 in Supplemental Information for the original ChromoPainter paper (Lawson et al., 2012) for motivation behind our parameterization, which is as follows:

$$\rho = \exp(-s \times cM^\gamma) - 1$$

References

Lawson, D. J., Hellenthal, G., Myers, S., & Falush, D. (2012). Inference of population structure using dense haplotype data. PLoS genetics, 8(1).

See also

Examples

if (FALSE) {
# Load the mini example recombination map from package:
small.map <- read.table(system.file("small_example/small.map", package = "kalis"), header = TRUE)
# Or, from the package built-in dataset
data(SmallMap)

# Compute the recombination probabilities
rho <- CalcRho(diff(SmallMap))
}