Copy Forward/Backward tables

Copies the contents of one forward/backward table into another.

Usage

CopyTable(to, from)

Arguments

to

a kalisForwardTable or kalisBackwardTable object which is to be copied into.

from

a kalisForwardTable or kalisBackwardTable object which is to be copied from.

Details

The core code in kalis operates on forward and backward tables at a very low level, both for speed (using low level CPU vector instructions) but also to avoid unnecessary memory copies since these tables will tend to be very large in serious genetics applications. As a result, if you attempt to copy a table in the standard idomatic way in R:

fwd2 <- fwd

then these two variables merely point to the same table: running the forward algorithm on fwd would result in fwd2 also changing.

This function is therefore designed to enable explicit copying of tables.

See also

Examples

# Examples
if (FALSE) {
# Create a forward table for the hidden Markov model incorporating all
# recipient and donor haplotypes
fwd <- MakeForwardTable()

# Propagate forward to variant 10:
Forward(fwd, pars, 10, nthreads = 8)

# This does **NOT** work as intended:
fwd2 <- fwd
Forward(fwd, pars, 20, nthreads = 8)

# Both tables are now at variant 20
fwd
fwd2

# Instead, to copy we create another table and use this function
fwd2 <- MakeForwardTable()
CopyTable(fwd2, fwd)
}