Distance matrix

Calculates a local distance matrix.

DistMat(fwd, bck)

Arguments

fwd
a forward table as returned by MakeForwardTable

bck
a backward table as returned by MakeBackwardTable

Value

Matrix of distances. The \((j,i)\)-th element of of the returned matrix is the inferred distance \(d_{ji}\) between haplotypes \(j\) and \(i\) at the current locus.

Details

This computes a local distance matrix based on the forward and backward tables at a certain locus. The forward and backward tables provided must be at the same locus \(l\). The \((j,i)\)-th element of of the returned matrix is the inferred distance \(d_{ji}\) between haplotypes \(j\) and \(i\) at the current locus, \(l\), of the two tables given the haplotypes observed (over the whole sequence).

In particular,

$$d_{ji} = -\frac{log(p_{ji}) + log(p_{ij})}{2}$$

where \(p_{ji}\) is the posterior marginal probability that \(j\) is coped by \(i\) at the current locus of the two tables, \(l\), given the haplotypes observed (over the whole sequence).

By convention, \(d_{ii} = 0\) for all \(i\).

See also

Examples

if (FALSE) {
# To get the distances at, say, locus 10 ...
Forward(fwd, pars, 10, nthreads = 8)
Backward(bck, pars, 10, nthreads = 8)
d <- DistMat(fwd, bck)
}