Distance matrix

Utility for calculating posterior marginal probabilities and distance matrices at, in between, or excluding variants.

DistMat(
  fwd,
  bck,
  standardize = FALSE,
  M = NULL,
  beta.theta.opts = NULL,
  nthreads = 1
)

Arguments

fwd
a forward table as returned by MakeForwardTable

bck
a backward table as returned by MakeBackwardTable

standardize
a logical; should the columns be centered and scaled to have unit variance

M
an pre-existing matrix into which to write the distances, can yield substantial speed up but requires special attention (see Details)

beta.theta.opts
a list; see Details.

nthreads
the number of CPU cores to use. By default no parallelism is used.

from_recipient
offset for distributed problems, see kalis distributed.

Value

Matrix of distances. The \((j,i)\)-th element of of the returned matrix is the inferred distance \(d_{ji}\) to haplotype \(j\) from haplotype \(i\) at the current locus. Each column encodes the output of an independent HMM: in column \(i\), haplotype \(i\) is taken as the observed recipient haplotype and painted as a mosaic of the other \(N-1\) haplotypes. Hence, the distances are asymmetric.

If calculating many distance matrices in succession, providing a pre-existing matrix M that can be updated in-place can drastically increase speed by eliminating the time needed for memory allocation. Be warned, since the matrix is updated in-place, if any other variables point to the same memory address, they will also be simultaneously overwritten. For example, writing

M <- matrix(0,nrow(fwd$alpha),ncol(fwd$alpha))
P <- M
DistMat(fwd,bck,M=M)

will update M and P simultaneously.

When provided, M must have dimensions matching that of fwd$alpha. Typically, that is simply \(N\) x \(N\) for \(N\) haplotypes. However, if kalis is being run distributed, M will be a \(N\) x \(R\) matrix where \(R\) is the number of recipient haplotypes on the given machine.

In order to obtain distance matrices between variants fwd$l and bck$l, then bck must be in beta.theta space. This allows the forward and backward tables to be transitioning both tables to some genomic position between fwd$l and bck$l. The precise recombination distance by which each table is propagated can be controlled by passing optional arguments in a list via beta.theta.opts. The recombination distances used can be specified in one of two ways.

Manually. In this case, beta.theta.opts is a list containing two named elements:

"rho.fwd" \(\in (0,1)\) specifies the transition probability \(rho\) for propagating the forward table

"rho.bck" \(\in (0,1)\) specifies the transition probability \(rho\) for propagating the backward table.

Implicitly. In this case, beta.theta.opts is a list containing two named elements:

"pars": a kalisParameters object that implicitly defines the recombination distance \(\rho^\star\) between fwd$l and bck$l

"bias" \(\in (0,1)\). The forward table is propagated a distance of bias\(\rho^\star\) and the backward table is propagated a distance of (1-bias)\(\rho^\star\).

Details

This computes a local probability or distance matrix based on the forward and backward tables at a certain locus. The default usage is provide forward and backward tables at the same locus \(l\). so that the \((j,i)\)-th element of of the returned matrix is the inferred distance \(d_{ji}\) between haplotypes \(j\) and \(i\) at the current locus, \(l\), of the two tables given the haplotypes observed (over the whole sequence).

In particular,

$$d_{ji} = -log(p_{ji})$$

where \(p_{ji}\) is the posterior marginal probability that \(j\) is coped by \(i\) at the current locus of the two tables, \(l\), given the haplotypes observed (over the whole sequence).

By convention, \(d_{ii} = 0\) for all \(i\).

This function also allows users to calculate distance matrices in between variants and also to calculate matrices that exclude a set of consecutive variants by passing a backward table in beta.theta space. If in beta.theta space, bck$l may be greater than but not equal to fwd$l. beta.theta.opts provides is required in this case to set how much of a recombination distance to propagate each matrix before combining them into distances. See Details below.

See also

Examples

if (FALSE) {
# To get the distances at, say, locus 10 ...
Forward(fwd, pars, 10, nthreads = 8)
Backward(bck, pars, 10, nthreads = 8)
d <- DistMat(fwd, bck)
}