Calculate a Distance Matrix from a Forward Table Object and a Backward Table Object

Provides an easy function for calculating a local distance matrix.

DistMat(fwd, bck)

Arguments

fwd
a forward table as returned by MakeForwardTable

bck
a backward table as returned by MakeBackwardTable

Value

matrix of distances

Details

The forward and backward tables must be at the same locus. The (i,j)-th element of of the returned matrix is the inferred distance d_(i,j) between haplotypes j copies i at locus fwd$l=bck$l given the haplotypes observed (from locus 1 to L).

d_(i,j) = -( log(p_(i,j)) + log(p_(j,i)) ) / 2 where p_(i,j) is the posterior marginal probability that j copies i at locus fwd$l=bck$l given the haplotypes observed (from locus 1 to L).

By convention, d_(i,i) = 0 for all i.

See also

Examples

DistMat(fwd,bck)

#> Error in DistMat(fwd, bck): object 'fwd' not found