Propagate a forward table

Propagates a kalisForwardTable to a downstream variant. The table is updated in-place.

Forward(
  fwd,
  pars,
  t = fwd$l + 1,
  nthreads = parallel::detectCores(logical = FALSE)
)

Arguments

fwd
a kalisForwardTable object, as returned by MakeForwardTable.

pars
a kalisParameters object, as returned by Parameters.

t
a variant to move the forward table to. Must be greater than or equal to current variant of fwd. By default, it simply advances to the next variant downstream.

nthreads
the number of CPU cores to use. By default uses the parallel package to detect the number of physical cores.

Value

There is nothing returned. For performance reasons, fwd is updated in-place.

Details

Forward implements the forward algorithm to advance the Li and Stephens rescaled hidden Markov model forward probabilities to a new target variant. Naturally, this can only propagate a table to variants downstream of its current position.

For mathematical details please see Section 2 of the kalis paper (TODO: ref). Note that the precise formulation of the forward equation is determined by whether the flag use.spiedel is set in the parameters provided in pars.

See also

Examples

if (FALSE) {
# Load the toy haplotype example and set toy parameters
CacheHaplotypes(SmallHaplotypes)
pars <- Parameters(rho = CalcRho(cM = SmallMap))

# Create a forward table for the hidden Markov model incorporating all
# recipient and donor haplotypes
fwd <- MakeForwardTable(pars)

# Create a forward table for the hidden Markov model incorporating only
# recipient haplotypes 100 to 200 (inclusive) and all donor haplotypes.
fwd <- MakeForwardTable(pars, 100, 200)

# This table is uninitialised, but ready to pass to the Forward function
# which will trigger initialisation and propagation.
# For example, initialise and propagate forward to the 10th variant:
Forward(fwd, pars, 10, nthreads = 8)
print(fwd)
}