Create a kalisBackwardTable

Allocates the memory for and initializes a backward table.

MakeBackwardTable(pars, from_recipient = 1, to_recipient = Inf)

Arguments

pars
a kalisParameters object specifying the genetics parameters to be associated with this forward table. These parameters can be set up by using the Parameters function.

from_recipient
first recipient haplotype if creating a partial forward table. By default includes from the first recipient haplotype.

to_recipient
last recipient haplotype if creating a partial forward table. By default includes to the last recipient haplotype.

Value

A specialized list of class kalisBackwardTable. The elements of the backward table list are:

l

denotes the current locus position.

beta

is a matrix of rescaled backward probabilities under the Li & Stephens HMM. Each column of beta corresponds to an independent HMM such that \(\beta^l_{ji}\) is proportional to the probability of observing haplotype \(i\) from locus \(l+1\) up through locus \(L\) given that haplotype \(j\) is copied by haplotype \(i\) at locus \(l\).

beta.g

is a vector containing scaling constants needed to continue propagating the HMM (please see kalis paper for details).

A kalisBackwardTable also carries with it a checksum key for the parameters pars it was provided. If one attempts to interact a kalisBackwardTable with a kalisForwardTable with mismatched parameters, an error will be thrown.

Details

MakeBackwardTable returns a kalisBackwardTable object appropriate for a given set of haplotypes (that must have already been cached by CacheHaplotypes) and a given set of Li and Stephens hidden Markov model parameters created by Parameters. The returned kalisBackwardTable is initialized at at the end of the cached haplotypes (technically bck$l=2,147,483,647 for computational reasons) and is ready to be propagated to a given target locus with the function Backward.

Since there is an independent hidden Markov model run for each recipient haplotype, it is possible to create a partial backward table object which corresponds to a subset of recipients using the from_recipient and to_recipient arguments.

See also

Examples

if (FALSE) {
# This code assumes you have already created the parameter set in a variable
# called pars

# Create a backward table for the hidden Markov model incorporating all
# recipient and donor haplotypes
bck <- MakeBackwardTable(pars)

# Create a backward table for the hidden Markov model incorporating only
# recipient haplotypes 100 to 200 (inclusive) and all donor haplotypes.
bck <- MakeBackwardTable(pars, 100, 200)

# This table is uninitialised, but ready to pass to the Backward function
# which will trigger initialisation and propagation from the last locus.
# For example, initialise and propagate backward to locus 10:
Backward(bck, pars, 10, nthreads = 8)
}