Construct parameter specification

Specify a parameter set to be used for a particular Li and Stephens hidden Markov model run.

Parameters(
  rho = rep(0, get("L", envir = pkgVars) - 1),
  mu = 1e-08,
  Pi = NULL,
  check.rho = TRUE
)

Arguments

rho
recombination probability vector (must be L-1 long). See CalcRho for assistance constructing this from standard genetics parameters.

mu
a scalar (for uniform) or vector (for varying) mutation costs.

Pi
leaving the default of uniform copying probabilities is recommended for computational efficiency. If desired, a full matrix of background copying probabilities can be provided, such that the (j,i)-th element is the background probability that j is copied by i. Hence, (a) the diagonal must be zero; and (b) the columns of Pi must sum to 1.

check.rho
if TRUE, a check that rho is within machine precision will be performed. If you have created rho using CalcRho with floor=TRUE then this will be satisfied automatically.

Value

A kalisParameters object, suitable for use to create the standard forward and backward recursion tables with MakeForwardTable and MakeBackwardTable. Note you will also need to provide this parameters object when propagating those tables using either Forward or Backward.

Details

There are 3 parameters which must be specified before the forward or backward equations in the Li and Stephens hidden Markov model can be run. These are the vector of recombination probabilities, the mutation costs, and the prior copying probabilities.

Recombination probabilities, rho

This is a vector parameter which must have length L-1, where L is the length (number of loci) of the haplotype sequences which have been loaded into the memory cache (using CacheHaplotypes). Note that element i of this vector should be the recombination probability between locus i and i+1.

There is a utility function, CalcRho, to assist with creating this recombination vector based on the more standard genetic parameters of Morgan distances, effective population size and a scalar power.

By default, the recombination probabilities are set to zero everywhere.

Mutation costs, mu

The mutation costs may be specified either as uniform across the whole length of the haplotypes (by providing a single scalar value), or may be varying at each locus (by providing a vector of length equal to the length of the haplotype sequences which have been loaded into the memory cache).

By default, mutation costs are set to \(10^{-8}\).

Copying probabilities, Pi

The original Li and Stephens model assumed that each haplotype has an equal prior probability of copying from any other. However, in the spirit of ChromoPainter (Lawson et al., 2012) we allow a matrix of prior copying probabilities.

The copying probabilities may be specified as a matrix of size \(N \times N\) (where \(N\) is the number of haplotypes which have been loaded into the memory cache), or for uniform copying probabilities need not be specified (resulting in copying probability \(\frac{1}{N-1}\) everywhere). Note that the diagonal must by definition be zero and columns must sum to one.

Note that there is a computational cost associated with non-uniform copying probabilities, so absent a strong motivation it is recommended to leave the default of uniform probabilities (NB do not specify a uniform matrix which would incur a computation cost too).

References

Lawson, D. J., Hellenthal, G., Myers, S., & Falush, D. (2012). Inference of population structure using dense haplotype data. PLoS genetics, 8(1).

See also

Examples

if (FALSE) {
# To use all the defaults
pars <- Parameters()

# To use indirect genetics parameters which you have already specified in
# morgan.dist, Ne and gamma variables to specify rho, leaving mu and Pi as
# defaults
pars <- Parameters(CalcRho(morgan.dist, Ne, gamma))

# Then use to create a table (eg for forward) ...
fwd <- MakeForwardTable(pars)
# ... and to propagate it
Forward(fwd, pars, 10, nthreads = 8)
}